+++ baseplate/experiments/providers/single_variant.py
@@ -59,9 +59,9 @@ class SingleVariantExperiment(SimpleExperiment):
                           any of the variants
         """
 
-        if bucket < (self.variants[0]["size"] * self.num_buckets):
+        if bucket < int(self.variants[0]["size"] * self.num_buckets):
             return self.variants[0]["name"]
-        elif bucket >= (self.num_buckets
+        elif bucket >= int(self.num_buckets
                 - (self.variants[1]["size"] * self.num_buckets)):
             return self.variants[1]["name"]
